follow pattern of 1) fetch 2) packages 3) generate script
move electrumx requirements workaround to the appropriate location (not setup.py!)
flesh out body of controller.py five main entry point functions and move subordinate callees to the appropriate class (resetters, stoppers etc.)
