+++ electrumsv-sdk/electrumsv_sdk/reset.py
@@ -2,18 +2,19 @@ import logging
 import os
 import shutil
 import subprocess
+import sys
 import time
 from pathlib import Path
 
 from electrumsv_node import electrumsv_node
 
-from .utils import logger
+from .constants import DEFAULT_ID_NODE, DEFAULT_ID_ELECTRUMX
 from .stoppers import Stoppers
-from .components import ComponentOptions, ComponentName
+from .components import ComponentOptions, ComponentName, ComponentStore, ComponentType
 from .installers import Installers
 from .starters import Starters
 
-logger = logging.getLogger("main")
+logger = logging.getLogger("resetters")
 orm_logger = logging.getLogger("peewee")
 orm_logger.setLevel(logging.WARNING)
 
@@ -24,6 +25,7 @@ class Resetters:
         self.starters = Starters(self.app_state)
         self.stoppers = Stoppers(self.app_state)
         self.installers = Installers(self.app_state)
+        self.component_store = ComponentStore(self.app_state)
 
     def normalize_wallet_name(self, wallet_name: str):
         if wallet_name is not None:
@@ -115,3 +117,53 @@ class Resetters:
         self.delete_wallet()
         self.create_wallet()
         logger.debug("Reset of RegTest electrumsv wallet completed successfully")
+
+    def reset_component_by_id(self, component_id):
+        if self.app_state.start_options[ComponentOptions.ID] != "":
+            if len(self.app_state.reset_set) != 0:
+                logger.debug(f"The '--id' flag is specified "
+                             f"- ignoring the component type(s) {self.app_state.reset_set}")
+
+        component_data = self.component_store.component_data_by_id(component_id)
+        if component_data == {}:
+            sys.exit(1)
+
+        if component_data.get('process_type') == ComponentType.NODE:
+            logger.debug(f"There is only one 'id' for this component type - resetting the "
+                         f"default node id={DEFAULT_ID_NODE}")
+            self.reset_node()
+
+        elif component_data.get('process_type') == ComponentType.ELECTRUMX:
+            logger.debug(f"There is only one 'id' for this component type - resetting the "
+                         f"default electrumx id={DEFAULT_ID_ELECTRUMX}")
+            self.reset_electrumx()
+
+        elif component_data.get('process_type') == ComponentType.ELECTRUMSV:
+            self.reset_electrumsv_wallet(component_id)
+
+        elif component_data.get('process_type') == ComponentType.INDEXER:
+            logger.debug(f"The Indexer component type is not supported at this time.")
+
+        elif component_data.get('process_type') == ComponentType.STATUS_MONITOR:
+            logger.error("resetting the status monitor is not supported at this time...")
+
+    def reset_component_by_type(self):
+        if ComponentName.NODE in self.app_state.reset_set or len(self.app_state.reset_set) == 0:
+            self.reset_node()
+
+        if ComponentName.ELECTRUMX in self.app_state.reset_set or len(
+                self.app_state.reset_set) == 0:
+            self.reset_electrumx()
+
+        if ComponentName.INDEXER in self.app_state.reset_set or len(
+                self.app_state.reset_set) == 0:
+            logger.error("resetting indexer is not supported at this time...")
+
+        if ComponentName.STATUS_MONITOR in self.app_state.reset_set \
+                or len(self.app_state.reset_set) == 0:
+            logger.error("resetting the status monitor is not supported at this time...")
+
+        if ComponentName.ELECTRUMSV in self.app_state.reset_set or len(
+                self.app_state.reset_set) == 0:
+            self.reset_electrumsv_wallet()
+            self.app_state.stop_set.add(ComponentName.ELECTRUMSV)
