--- electrumsv-sdk/requirements/requirements-electrumx.txt
@@ -1,4 +1,0 @@
-aiorpcX[ws]>=0.18.3,<0.19
-attrs
-pylru
-aiohttp>=3.3
