--- electrumsv-sdk/requirements/requirements-win32.txt
@@ -1,1 +1,0 @@
-plyvel-win32>=1.2.1
